Created Action Semicircle
Created an action circle
-It currently does nothing but increase to its limit (two seconds) and reset to zero
-Refactored so that it they can be created with one line of code (draw_act_circle function)

The idea is that, when an enemy attacks, one will spawn below the timeline at whatever point the timeline was at that moment. Then the player must dodge (press a button) when the green inner slice passes the yellow highlight.
None of that is currently implemented.