Enemy Attack
Enemy will now register attacks if the player is in range, at the interval of their attack speed.
The attack will have the yellow (target) slice for the action circle appear in a random position, but at a set size.
These action circles will appear with their center on where the bar was on the attack.
The circles will also disappear when they have filled up .