Keyboard Controls
Added gamepad slot and description to the debug UI

Generalized debug UI names

Added (admitedly terrible) keyboard controls

Debug UI modified to show when a button is pressed (for some inputs). Things in brackets show when the button is pressed. It's only there for that one frame, so it will only be a flicker.

Allowed player to finish moving to their next position (after inputting a move command) during the enemy turn. Before, the player would stop moving part way if it became the enemy turn during their move to another square.