A new switch and button prompts
Removed background

Added a new type of wall and switch as well as a new part of the test dungeon to test it

Discovered a bug when the player switch direction mid move. I can no longer reproduce it, so I believe that I fixed it.

Created an empty room that will become the tutorial

Test enemies now have a shade showing what color they attack on.

Created button prompts.
-Switches now show a button prompt when the player is able to toggle them.