Added tutorial and new triggers
Built more of the tutorial, including a new, tutorial specific enemy.

Added enemy death triggers

Added a reusability option to fake walls.

Added floor triggers (which can be invisible to work as area triggers).

Fixed a bug on horizontal walls where the walls (that become the sides of the hall) weren't visible.

Made starting room the tutorial area

Reaching the current end of the tutorial will teleport the player to the test dungeon