More saves
Solved the bug of the player being unable to unpause while in the test dungeon.
   -The issue was that a new master object was being created, which was messing on my pause logic, which is based on having a single, reliable object. I tracked it down, and evidentally I still had a master object placed in the room. Removing it seems to have fixed that issue.

Now enemy attack (peices) are saved an loaded for pauses.

Simplified a script, as I wasn't going to use it with two arguments I gave it.

Doors and their toggles are now saved and loaded properly on pause and unpause. (This breaks the other puzzle elements though)

Fixed memory leaks relating to the creation of my ds_structures (I wasn't deleting them after their usage).