Switch and Options Menu
Updated switch sprite

Added Fog of War option

Discovered a bug with FoV/FoW where it wouldn't penetrate invisible switches
   -I rectified this by making it so the FoV is only stopped by walls, and also by making the doors into a child wall.

The trigger to turn on the FoW in the tutorial now forces it on, as opposed to toggling it.

If the player's actions were disabled they are kept disabled on reload. Otherwise, it is reset to 0 (the start).

Tried to fix a bug where the player would teleport to the wrong place and then slide to a wrong place upon loading into a new room. (I haven't been able to replicate it, so I think it's fixed).

Removed the ability to change zoom levels

Added a part of the options menu for key binding  (currently no functionality and way out except to unpause).