More options
Fixed a bug where the player was not automatically moving when holding a direction.

Completed moving keyboard inputs to another level of abstraction

Added the ability to configure the look direction keyboard key bindings.

Added a new "column" to the keyboard key binding screen (although there is nothing there)

Added the ability to loop around vertically and horizontally on keyboard keybinding screen.