New combos
The center of the center circle in the state machine shows show far down the combo you are.

Removed some commented out code in the state machine

Added a new continuatino combo to the light physical combo, the light physical combo 2. (Light physical combo 1 is activated by 0, 0, 0. Light physical combo 2 is an additional 1, 0; meaning that it is activated by 0, 0, 0, 1, 0). It attacks only directly in front of the player, just like light physical combo 1.

Added the medium magic combo
   -Activated by 1, 1, 0, 1, 0
   -Uses a bonus damage multiplier of 1.6
   -Attacks two spots directly in front of the player (i.e., in front of the player, and behind that)

Balance changes
   -Nerfed the player's attack from 100 to 70 (remember that damage is not a linear damage calcuation)
   -Boosted the attack bonus multiplier of the light physical combo 2 and light magic combo. They went from 1 and 1.3 to 1.2 and 1.35 respectively.

Fixed a bug that caused a crash when trying to use the light magical combo 1