Disable-able abilities
The state machine is once again disabled upon starting the tutorial.

The state machine now has a separate variabled to keep track of if it is enabled or not. This means that you can no longer forcably reenable the state machine by pressing an action button.

State machine circles now have a white outline, allowing them to be seen easier (or at all, in the case of being on a totally black background)

Time can now be disabled

Time is now disabled upon started the tutorial, and is not enabled until leaving the first "room" of the tutorial

Dodging can now be disabld

Dodging is now disabled upon starting the tutorial and is now reactivated at the same time as attacking.

Toggle objects can no longer be made into a "child" of toggle objects