Peice fix
Reworked inventory system and keys so they will now be separate based on room/dungeon

Keys are now saved and loaded correctly on pause

Began adding a miniboxx

Increased opacity of rhythm segment of UI (0.2 -> 0.4)

Found a bug where the player was able to walk through locked doors
     (Not yet fixed)

Adjusted the dodge/rhythm rectangles so that the middle corresponds with their exact position

Added red line for center of rhythm peices to debug UI

Fixed peice spawning and movement so it will now work respective of how fast the peices will move.

Added a temporary extra of the miniboss in the spawn room for quick testing.