Particle Effects
When a rhythm peice is destroyed (and doesn't hurt the player), it evaporates into particles.

When the player does damage to an enemy, particles are spawned from the enemy in the direction opposite from where the player is (as if emenating from the player). The number is proportional to the damage done (i.e., 1 particle for every 10 damage done).